Added the player plane to the input
Now the input has size of `NxN`, doesn't
have a plane for the wall, but got two planes
to tell which player makes the first move.

The output now is whether the group lives
or dies if the given player makes the first
move.

Samples are also egenrated differently:
instead of picking moves that change the
safety status, the generator now picks a few
random moves for the player that's losing.